Update patch set 1
Patch Set 1: Do not submit

The 'SD card' in Nexus devices is actually part of /data. You can't allow physical block access to /data for both the system and USB; data loss will be the consequence. You only can do UMS on separate partitions and/or disks.

Exactly that is the reason why AOSP implemented MTP/PTP in the first place.

Patch-set: 1