Update patch set 6
Patch Set 6:

IMHO, when you talk about UX you talk about random user trying the device out of the box and what the logic imposes that he would expect. Under that vision, I think user would expect a confortable and soft feedback and not a little shock on his finger. That said, I see zero benefits on increasing haptic feedback by default. Google's default feels nice and probably match to the major part of userbase preference.

My number one option is deliver a nice experience out of the box and therefore not deploy a setting if there is not a big enough demand for it that justifys the adding. But between raising the default and have the setting I would go for the setting. I wouldn't mess with a stablished default that impact every user unless there is a clear benefit.

Patch-set: 6