Update patch set 6
Patch Set 6: Do not submit

I was a huge disapprover of the addition of GNexusParts and remain staunchly opposed to having this on the Nexus 4. Furthermore, this is even more pointless on the N4 than it was the GN.

On the Galaxy Nexus, the color saturation was out of whack to due AMOLED and the vibration motor wasn't nearly as nice as the linear oscillating one found in the Nexus 4. 

I will reverse this -2 if and only if Arcee disagrees with me on this, but I am fairly certain he won't.

Patch-set: 6