Update patch set 6
Patch Set 6:

Danny, Drew: Just to be clear, this is unrelated to the implementation. But I have stated on multiple occasions that I do not think these things (hardware-specific register fiddling) belong in the base OS. Not for the N4, and not for any device.
Not only they apply to a minority of users, but above all there's plenty of third-party applications that do this kind of... adjustment. Even if the N4 needs one specific to it, it can still be written and made available separately.

Patch-set: 6