Update patch set 6
Patch Set 6:

I agree to Miika Koskinen, there is really no need for the color adjustment settings because they are already very well balanced on the N4. The vibration adjustment is a useful feature in my eyes and was well coded by Danny so I think it‘s a good idea to implement it in the sound settings.

Patch-set: 6