Update patch set 6
Patch Set 6:

color calibration is a 'nice to have' feature. I tried this with another kernel and faux clock application. Also I lowered vibration strength because the default was to 'heavy'.

My suggestion is to create a free CM based add-on application that allows to change these settings for supported devices/kernels.

But there are some risks for users. I saw such problems at SGS2 that was able to change screen tint, etc.

Patch-set: 6