Update patch set 1
Patch Set 1: Code-Review-1

The advanced reboot menu is enabled through developer options, so it's not meant to be used by normal users. Removing it doesn't have any sense to me. It's enabled since forever for all of the devices that support it.

If you are afraid of bricking the device in bootloader mode, you shouldn't have messed with developer settings in the first place :)

Patch-set: 1
Label: Code-Review=-1