Update patch set 2
Patch Set 2:

Then yea, I'd rather wait for 5.1. 

Master frequently has items that are withheld from tagged releases (for unknown reasons) and we'll get the rest of the changes as part of the 5.1 merge into this tree. 

If there is something in this series that addresses an item that we see now (clang issues or SE denials) then we can take those piecemeal, but tracking master shouldn't be one of our go-to's.

Patch-set: 2
Reviewer: Abhisek Devkota <848@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0