Update patch set 1
Patch Set 1:

You aren't misreading the intention of the patch.

This write was meant to enable frequency boosts on thread migrations in the ondemand governor. Since Lollipop however, the file being written to is no longer there and the correct file is now /dev/cpuctl/cpu.notify_on_migrate.

The appropriate patch then would be to write to the new file. However, on flo, interactive governor is now the default and we no longer need the write (since it was meant for ondemand governor). Thus, I removed it the entire write instead (this patch).

Mako still uses ondemand governor so the appropriate patch would be to write to the new file: /dev/cpuctl/cpu.notify_on_migrate.

Patch-set: 1